requirements.txt
kafka-python==2.0.2
asyncpg==0.23.0
